Add enums for user types, work types, days, and payment statuses; update DTOs and repository for improved data handling; modify main.dart for welcome page import; adjust build.gradle.kts for NDK version; add new dependencies in pubspec.yaml; enhance application properties for JPA; implement BusinessException class; and create new DTOs for login and various entities.

+++ backend/src/main/java/com/gestiongimnasio/backend/dto/post/InscripcionPostDTO.java
@@ -1,14 +1,8 @@
 package com.gestiongimnasio.backend.dto.post;
 
 import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
 
 @Data
 public class InscripcionPostDTO {
@@ -16,7 +10,7 @@ public class InscripcionPostDTO {
     private Long clienteId;
 
     @NotNull
-    private Integer claseId;
+    private Long claseId;
 
-    private Integer pagoId;
+    private Long pagoId;
 }
